refactor: rework structure of app-data folder
