+++ src/app-data/__tests__/getGCMSMeasurement.test.ts
@@ -1,8 +1,7 @@
 import { expect, test } from 'vitest';
 
 import { getTestFileCollection } from '../../utils/test-utils';
-import { loadMeasurements } from '../loadMeasurements';
-import { cdfLoader } from '../loaders/cdfLoader';
+import { loadMeasurements, cdfLoader } from '../index';
 
 const fileCollection = await getTestFileCollection('cdf');
