+++ src/app-data/__tests__/loadMeasurements.test.ts
@@ -1,11 +1,13 @@
 import { test, expect } from 'vitest';
 
 import { getTestFileCollection } from '../../utils/test-utils';
-import { getIRAutoPeakPickingEnhancer } from '../kinds/ir/irAutoPeakPickingEnhancer';
-import { irMeasurementEnhancer } from '../kinds/ir/irMeasurementEnhancer';
-import { loadMeasurements } from '../loadMeasurements';
-import { jcampLoader } from '../loaders/jcampLoader';
-import { wdfLoader } from '../loaders/wdfLoader';
+import {
+  loadMeasurements,
+  jcampLoader,
+  wdfLoader,
+  getIRAutoPeakPickingEnhancer,
+  irMeasurementEnhancer,
+} from '../index';
 
 const loaders = [jcampLoader, wdfLoader];
 const enhancers = {
