+++ src/app-data/kinds/index.ts
@@ -1,0 +1,1 @@
+export * from './ir/index';
