+++ src/app-data/kinds/ir/index.ts
@@ -1,0 +1,4 @@
+export * from './irAutoPeakPickingEnhancer';
+export * from './IRMeasurement';
+export * from './irMeasurementEnhancer';
+export * from './IRPeak';
