+++ src/app-data/loaders/proprietary/agilent/cary500Loader.ts
@@ -1,9 +1,12 @@
 import type { FileCollectionItem, FileCollection } from 'filelist-utils';
 
-import type { Measurements } from '../../../DataState';
-import type { MeasurementBase } from '../../../MeasurementBase';
-import { getMeasurementInfoFromFile } from '../../utility/getMeasurementInfoFromFile';
-import { ParserLog, createLogEntry } from '../../utility/parserLog';
+import {
+  Measurements,
+  MeasurementBase,
+  getMeasurementInfoFromFile,
+  ParserLog,
+  createLogEntry,
+} from '../../../index';
 
 export async function cary500Loader(
   fileCollection: FileCollection,
