+++ src/app-data/loaders/spcLoader.ts
@@ -2,8 +2,7 @@ import type { FileCollection } from 'filelist-utils';
 import { parse, guessSpectraType } from 'spc-parser';
 
 import { assert } from '../../utils/assert';
-import type { MeasurementKind, Measurements } from '../DataState';
-import type { MeasurementBase } from '../MeasurementBase';
+import type { MeasurementKind, Measurements, MeasurementBase } from '../index';
 
 import { getMeasurementInfoFromFile } from './utility/getMeasurementInfoFromFile';
 import { ParserLog, createLogEntry } from './utility/parserLog';
