+++ src/app-data/loaders/wdfLoader.ts
@@ -1,8 +1,7 @@
 import type { FileCollection } from 'filelist-utils';
 import { parse } from 'wdf-parser';
 
-import type { Measurements } from '../DataState';
-import type { MeasurementBase } from '../MeasurementBase';
+import type { Measurements, MeasurementBase } from '../index';
 
 import { getMeasurementInfoFromFile } from './utility/getMeasurementInfoFromFile';
 import { ParserLog, createLogEntry } from './utility/parserLog';
