+++ src/app-data/state/data/kinds.ts
@@ -1,0 +1,16 @@
+import type { MeasurementKind } from './DataState';
+
+export const kindLabels: Record<MeasurementKind, string> = {
+  ir: 'IR',
+  iv: 'IV',
+  raman: 'Raman',
+  uv: 'UV',
+  uvvis: 'UV-VIS',
+  mass: 'Mass',
+  gclc: 'GC/LC',
+  gclcms: 'GC/LC MS',
+  nmr: 'NMR',
+  other: 'Other',
+};
+
+export const measurementKinds = Object.keys(kindLabels) as MeasurementKind[];
