+++ src/app-data/state/settings/index.ts
@@ -1,0 +1,1 @@
+export type DummySettings = string;
