+++ src/app-data/state/view/index.ts
@@ -1,0 +1,1 @@
+export type DummyView = string;
