+++ src/app/index.ts
@@ -1,4 +1,3 @@
 export * from './explorer/index';
-export * from './kinds/ir/index';
-export * from './kinds/mass/index';
+export * from './kinds/index';
 export * from './panels/index';
