+++ src/app/kinds/index.ts
@@ -1,0 +1,2 @@
+export * from './ir';
+export * from './mass';
