+++ src/app/panels/MeasurementPanel.tsx
@@ -1,5 +1,5 @@
-import { MeasurementDisplay, useAppDispatch } from '../../app-data/appState';
-import type { MeasurementKindAndId } from '../../app-data/data.helpers';
+import { MeasurementDisplay, useAppDispatch } from '../../app-data/index';
+import type { MeasurementKindAndId } from '../../app-data/index';
 import { ColorPicker } from '../../components/index';
 
 export function MeasurementPanel(props: {
