+++ src/app/panels/index.ts
@@ -1,3 +1,4 @@
 export * from './MeasurementInfoPanel';
+export * from './MeasurementPanel';
 export * from './MeasurementsPanel';
 export * from './SignalProcessingPanel';
