+++ src/pages/demo/MainLayout.tsx
@@ -22,8 +22,8 @@ import {
   MeasurementExplorer,
   MeasurementInfoPanel,
   MeasurementsPanel,
+  MeasurementPanel,
 } from '../../app/index';
-import { MeasurementPanel } from '../../app/panels/MeasurementPanel';
 import {
   Accordion,
   DropZoneContainer,
